No changes. Just committing to provide some info
Contrary to what I thought earlier, it's still not working correctly (I
think I see patterns that shouldn't exist in the game of life [huge
chunks of living cells which don't respect the Game of Life rules). I
might be mistaken however (because the terminating pattern seems to be
correct - as compared to the iterative version).